MessagingMenuV3@blub: Update Application Support, added support for unsupported and updated applications
Current version of Messaging Menu V3 is missing support for newer messaging applications, and older applications that used to be supported no longer show up in the menu due to a change in the way GNOME identifies its applications. I've kept in the older geary and evolution ID's as well as the newer org.gnome.* versions in case this is used on older systems that might still support the older applications.